refactor(frontend): 重构前端架构，提取 composables 与组件拆分
- 新增 composables 层：useAppChat、useChatHistory、useCodeGeneration 等
- 拆分 AppChatPage 为 5 个子组件，主页面从 1080 行降至约 150 行
- 管理后台页面统一使用 useTablePage 复用分页与搜索逻辑
- 路由改为懒加载，统一 import 规范（移除 .ts 后缀）
- 修复登录 redirect 跳转、补充对话管理菜单项
- 补充缺失的 SVG 资源与 dayjs、@ant-design/icons-vue 依赖
- 改进 Pinia store 类型安全
