+++ yu-ai-code-mother-frontend/package.json
@@ -15,9 +15,10 @@
     "format": "prettier --write src/"
   },
   "dependencies": {
+    "@ant-design/icons-vue": "^7.0.1",
     "ant-design-vue": "^4.2.6",
     "axios": "^1.11.0",
-    "@types/markdown-it": "^14.1.2",
+    "dayjs": "^1.11.13",
     "highlight.js": "^11.11.1",
     "markdown-it": "^14.1.0",
     "pinia": "^3.0.3",
@@ -26,6 +27,7 @@
   },
   "devDependencies": {
     "@tsconfig/node22": "^22.0.2",
+    "@types/markdown-it": "^14.1.2",
     "@types/node": "^22.15.32",
     "@umijs/openapi": "^1.13.15",
     "@vitejs/plugin-vue": "^6.0.0",
