+++ yu-ai-code-mother-frontend/src/constants/menu.ts
@@ -1,0 +1,35 @@
+import { h } from 'vue'
+import { HomeOutlined } from '@ant-design/icons-vue'
+import type { MenuProps } from 'ant-design-vue'
+
+export const ADMIN_MENU_ITEMS: MenuProps['items'] = [
+  {
+    key: '/admin/userManage',
+    label: '用户管理',
+    title: '用户管理',
+  },
+  {
+    key: '/admin/appManage',
+    label: '应用管理',
+    title: '应用管理',
+  },
+  {
+    key: '/admin/chatManage',
+    label: '对话管理',
+    title: '对话管理',
+  },
+]
+
+export const PUBLIC_MENU_ITEMS: MenuProps['items'] = [
+  {
+    key: '/',
+    icon: () => h(HomeOutlined),
+    label: '主页',
+    title: '主页',
+  },
+  {
+    key: 'others',
+    label: h('a', { href: 'https://www.codefather.cn', target: '_blank' }, '编程导航'),
+    title: '编程导航',
+  },
+]
