+++ yu-ai-code-mother-frontend/src/pages/admin/AppManagePage.vue
@@ -1,6 +1,5 @@
 <template>
   <div id="appManagePage">
-    <!-- 搜索表单 -->
     <a-form layout="inline" :model="searchParams" @finish="doSearch">
       <a-form-item label="应用名称">
         <a-input v-model:value="searchParams.appName" placeholder="输入应用名称" />
@@ -30,13 +29,13 @@
     </a-form>
     <a-divider />
 
-    <!-- 表格 -->
     <a-table
       :columns="columns"
       :data-source="data"
       :pagination="pagination"
-      @change="doTableChange"
+      :loading="loading"
       :scroll="{ x: 1200 }"
+      @change="onTableChange"
     >
       <template #bodyCell="{ column, record }">
         <template v-if="column.dataIndex === 'cover'">
@@ -56,9 +55,7 @@
           <span v-else>{{ record.priority || 0 }}</span>
         </template>
         <template v-else-if="column.dataIndex === 'deployedTime'">
-          <span v-if="record.deployedTime">
-            {{ formatTime(record.deployedTime) }}
-          </span>
+          <span v-if="record.deployedTime">{{ formatTime(record.deployedTime) }}</span>
           <span v-else class="text-gray">未部署</span>
         </template>
         <template v-else-if="column.dataIndex === 'createTime'">
@@ -69,12 +66,12 @@
         </template>
         <template v-else-if="column.key === 'action'">
           <a-space>
-            <a-button type="primary" size="small" @click="editApp(record)"> 编辑 </a-button>
+            <a-button type="primary" size="small" @click="editApp(record)">编辑</a-button>
             <a-button
               type="default"
               size="small"
-              @click="toggleFeatured(record)"
               :class="{ 'featured-btn': record.priority === 99 }"
+              @click="toggleFeatured(record)"
             >
               {{ record.priority === 99 ? '取消精选' : '精选' }}
             </a-button>
@@ -88,11 +85,11 @@
   </div>
 </template>
 
-<script lang="ts" setup>
-import { computed, onMounted, reactive, ref } from 'vue'
+<script setup lang="ts">
 import { useRouter } from 'vue-router'
 import { message } from 'ant-design-vue'
 import { listAppVoByPageByAdmin, deleteAppByAdmin, updateAppByAdmin } from '@/api/appController'
+import { useTablePage } from '@/composables/useTablePage'
 import { CODE_GEN_TYPE_OPTIONS, formatCodeGenType } from '@/utils/codeGenTypes'
 import { formatTime } from '@/utils/time'
 import UserInfo from '@/components/UserInfo.vue'
@@ -100,138 +97,42 @@ import UserInfo from '@/components/UserInfo.vue'
 const router = useRouter()
 
 const columns = [
-  {
-    title: 'ID',
-    dataIndex: 'id',
-    width: 80,
-    fixed: 'left',
-  },
-  {
-    title: '应用名称',
-    dataIndex: 'appName',
-    width: 150,
-  },
-  {
-    title: '封面',
-    dataIndex: 'cover',
-    width: 100,
-  },
-  {
-    title: '初始提示词',
-    dataIndex: 'initPrompt',
-    width: 200,
-  },
-  {
-    title: '生成类型',
-    dataIndex: 'codeGenType',
-    width: 100,
-  },
-  {
-    title: '优先级',
-    dataIndex: 'priority',
-    width: 80,
-  },
-  {
-    title: '部署时间',
-    dataIndex: 'deployedTime',
-    width: 160,
-  },
-  {
-    title: '创建者',
-    dataIndex: 'user',
-    width: 120,
-  },
-  {
-    title: '创建时间',
-    dataIndex: 'createTime',
-    width: 160,
-  },
-  {
-    title: '操作',
-    key: 'action',
-    width: 200,
-    fixed: 'right',
-  },
+  { title: 'ID', dataIndex: 'id', width: 80, fixed: 'left' as const },
+  { title: '应用名称', dataIndex: 'appName', width: 150 },
+  { title: '封面', dataIndex: 'cover', width: 100 },
+  { title: '初始提示词', dataIndex: 'initPrompt', width: 200 },
+  { title: '生成类型', dataIndex: 'codeGenType', width: 100 },
+  { title: '优先级', dataIndex: 'priority', width: 80 },
+  { title: '部署时间', dataIndex: 'deployedTime', width: 160 },
+  { title: '创建者', dataIndex: 'user', width: 120 },
+  { title: '创建时间', dataIndex: 'createTime', width: 160 },
+  { title: '操作', key: 'action', width: 200, fixed: 'right' as const },
 ]
 
-// 数据
-const data = ref<API.AppVO[]>([])
-const total = ref(0)
-
-// 搜索条件
-const searchParams = reactive<API.AppQueryRequest>({
-  pageNum: 1,
-  pageSize: 10,
-})
-
-// 获取数据
-const fetchData = async () => {
-  try {
-    const res = await listAppVoByPageByAdmin({
-      ...searchParams,
-    })
-    if (res.data.data) {
-      data.value = res.data.data.records ?? []
-      total.value = res.data.data.totalRow ?? 0
-    } else {
+const { data, loading, searchParams, pagination, fetchData, doSearch, onTableChange } =
+  useTablePage<API.AppVO, API.AppQueryRequest>({
+    defaultParams: { pageNum: 1, pageSize: 10 },
+    fetchFn: async (params) => {
+      const res = await listAppVoByPageByAdmin(params)
+      if (res.data.data) {
+        return res.data.data
+      }
       message.error('获取数据失败，' + res.data.message)
-    }
-  } catch (error) {
-    console.error('获取数据失败：', error)
-    message.error('获取数据失败')
-  }
-}
+      return null
+    },
+  })
 
-// 页面加载时请求一次
-onMounted(() => {
-  fetchData()
-})
-
-// 分页参数
-const pagination = computed(() => {
-  return {
-    current: searchParams.pageNum ?? 1,
-    pageSize: searchParams.pageSize ?? 10,
-    total: total.value,
-    showSizeChanger: true,
-    showTotal: (total: number) => `共 ${total} 条`,
-  }
-})
-
-// 表格变化处理
-const doTableChange = (page: { current: number; pageSize: number }) => {
-  searchParams.pageNum = page.current
-  searchParams.pageSize = page.pageSize
-  fetchData()
-}
-
-// 搜索
-const doSearch = () => {
-  // 重置页码
-  searchParams.pageNum = 1
-  fetchData()
-}
-
-// 编辑应用
 const editApp = (app: API.AppVO) => {
   router.push(`/app/edit/${app.id}`)
 }
 
-// 切换精选状态
 const toggleFeatured = async (app: API.AppVO) => {
   if (!app.id) return
-
   const newPriority = app.priority === 99 ? 0 : 99
-
   try {
-    const res = await updateAppByAdmin({
-      id: app.id,
-      priority: newPriority,
-    })
-
+    const res = await updateAppByAdmin({ id: app.id, priority: newPriority })
     if (res.data.code === 0) {
       message.success(newPriority === 99 ? '已设为精选' : '已取消精选')
-      // 刷新数据
       fetchData()
     } else {
       message.error('操作失败：' + res.data.message)
@@ -242,15 +143,12 @@ const toggleFeatured = async (app: API.AppVO) => {
   }
 }
 
-// 删除应用
 const deleteApp = async (id: number | undefined) => {
   if (!id) return
-
   try {
     const res = await deleteAppByAdmin({ id })
     if (res.data.code === 0) {
       message.success('删除成功')
-      // 刷新数据
       fetchData()
     } else {
       message.error('删除失败：' + res.data.message)
