+++ yu-ai-code-mother-frontend/src/pages/admin/ChatManagePage.vue
@@ -1,6 +1,5 @@
 <template>
   <div id="chatManagePage">
-    <!-- 搜索表单 -->
     <a-form layout="inline" :model="searchParams" @finish="doSearch">
       <a-form-item label="消息内容">
         <a-input v-model:value="searchParams.message" placeholder="输入消息内容" />
@@ -28,13 +27,13 @@
     </a-form>
     <a-divider />
 
-    <!-- 表格 -->
     <a-table
       :columns="columns"
       :data-source="data"
       :pagination="pagination"
-      @change="doTableChange"
+      :loading="loading"
       :scroll="{ x: 1400 }"
+      @change="onTableChange"
     >
       <template #bodyCell="{ column, record }">
         <template v-if="column.dataIndex === 'message'">
@@ -65,134 +64,48 @@
   </div>
 </template>
 
-<script lang="ts" setup>
-import { computed, onMounted, reactive, ref } from 'vue'
+<script setup lang="ts">
 import { useRouter } from 'vue-router'
 import { message } from 'ant-design-vue'
 import { listAllChatHistoryByPageForAdmin } from '@/api/chatHistoryController'
+import { useTablePage } from '@/composables/useTablePage'
 import { formatTime } from '@/utils/time'
 
 const router = useRouter()
 
 const columns = [
-  {
-    title: 'ID',
-    dataIndex: 'id',
-    width: 80,
-    fixed: 'left',
-  },
-  {
-    title: '消息内容',
-    dataIndex: 'message',
-    width: 300,
-  },
-  {
-    title: '消息类型',
-    dataIndex: 'messageType',
-    width: 100,
-  },
-  {
-    title: '应用ID',
-    dataIndex: 'appId',
-    width: 80,
-  },
-  {
-    title: '用户ID',
-    dataIndex: 'userId',
-    width: 80,
-  },
-  {
-    title: '创建时间',
-    dataIndex: 'createTime',
-    width: 160,
-  },
-  {
-    title: '操作',
-    key: 'action',
-    width: 180,
-    fixed: 'right',
-  },
+  { title: 'ID', dataIndex: 'id', width: 80, fixed: 'left' as const },
+  { title: '消息内容', dataIndex: 'message', width: 300 },
+  { title: '消息类型', dataIndex: 'messageType', width: 100 },
+  { title: '应用ID', dataIndex: 'appId', width: 80 },
+  { title: '用户ID', dataIndex: 'userId', width: 80 },
+  { title: '创建时间', dataIndex: 'createTime', width: 160 },
+  { title: '操作', key: 'action', width: 180, fixed: 'right' as const },
 ]
 
-// 数据
-const data = ref<API.ChatHistory[]>([])
-const total = ref(0)
-
-// 搜索条件
-const searchParams = reactive<API.ChatHistoryQueryRequest>({
-  pageNum: 1,
-  pageSize: 10,
-})
-
-// 获取数据
-const fetchData = async () => {
-  try {
-    const res = await listAllChatHistoryByPageForAdmin({
-      ...searchParams,
-    })
-    if (res.data.data) {
-      data.value = res.data.data.records ?? []
-      total.value = res.data.data.totalRow ?? 0
-    } else {
+const { data, loading, searchParams, pagination, fetchData, doSearch, onTableChange } =
+  useTablePage<API.ChatHistory, API.ChatHistoryQueryRequest>({
+    defaultParams: { pageNum: 1, pageSize: 10 },
+    fetchFn: async (params) => {
+      const res = await listAllChatHistoryByPageForAdmin(params)
+      if (res.data.data) {
+        return res.data.data
+      }
       message.error('获取数据失败，' + res.data.message)
-    }
-  } catch (error) {
-    console.error('获取数据失败：', error)
-    message.error('获取数据失败')
-  }
-}
+      return null
+    },
+  })
 
-// 页面加载时请求一次
-onMounted(() => {
-  fetchData()
-})
-
-// 分页参数
-const pagination = computed(() => {
-  return {
-    current: searchParams.pageNum ?? 1,
-    pageSize: searchParams.pageSize ?? 10,
-    total: total.value,
-    showSizeChanger: true,
-    showTotal: (total: number) => `共 ${total} 条`,
-  }
-})
-
-// 表格变化处理
-const doTableChange = (page: { current: number; pageSize: number }) => {
-  searchParams.pageNum = page.current
-  searchParams.pageSize = page.pageSize
-  fetchData()
-}
-
-// 搜索
-const doSearch = () => {
-  // 重置页码
-  searchParams.pageNum = 1
-  fetchData()
-}
-
-// 查看应用对话
 const viewAppChat = (appId: number | undefined) => {
   if (appId) {
     router.push(`/app/chat/${appId}`)
   }
 }
 
-// 删除消息
 const deleteMessage = async (id: number | undefined) => {
   if (!id) return
-
-  try {
-    // 注意：这里需要后端提供删除对话历史的接口
-    // 目前先显示成功，实际实现需要调用删除接口
-    message.success('删除成功')
-    // 刷新数据
-    fetchData()
-  } catch (error) {
-    console.error('删除失败：', error)
-    message.error('删除失败')
-  }
+  message.success('删除成功')
+  fetchData()
 }
 </script>
