+++ yu-ai-code-mother-frontend/src/pages/admin/UserManagePage.vue
@@ -1,6 +1,5 @@
 <template>
   <div id="userManagePage">
-    <!-- 搜索表单 -->
     <a-form layout="inline" :model="searchParams" @finish="doSearch">
       <a-form-item label="账号">
         <a-input v-model:value="searchParams.userAccount" placeholder="输入账号" />
@@ -13,27 +12,24 @@
       </a-form-item>
     </a-form>
     <a-divider />
-    <!-- 表格 -->
     <a-table
       :columns="columns"
       :data-source="data"
       :pagination="pagination"
-      @change="doTableChange"
+      :loading="loading"
+      @change="onTableChange"
     >
       <template #bodyCell="{ column, record }">
         <template v-if="column.dataIndex === 'userAvatar'">
           <a-image :src="record.userAvatar" :width="120" />
         </template>
         <template v-else-if="column.dataIndex === 'userRole'">
-          <div v-if="record.userRole === 'admin'">
-            <a-tag color="green">管理员</a-tag>
-          </div>
-          <div v-else>
-            <a-tag color="blue">普通用户</a-tag>
-          </div>
+          <a-tag :color="record.userRole === 'admin' ? 'green' : 'blue'">
+            {{ record.userRole === 'admin' ? '管理员' : '普通用户' }}
+          </a-tag>
         </template>
         <template v-else-if="column.dataIndex === 'createTime'">
-          {{ dayjs(record.createTime).format('YYYY-MM-DD HH:mm:ss') }}
+          {{ formatTime(record.createTime) }}
         </template>
         <template v-else-if="column.key === 'action'">
           <a-button danger @click="doDelete(record.id)">删除</a-button>
@@ -42,114 +38,47 @@
     </a-table>
   </div>
 </template>
-<script lang="ts" setup>
-import { computed, onMounted, reactive, ref } from 'vue'
-import { deleteUser, listUserVoByPage } from '@/api/userController.ts'
+
+<script setup lang="ts">
 import { message } from 'ant-design-vue'
-import dayjs from 'dayjs'
+import { deleteUser, listUserVoByPage } from '@/api/userController'
+import { useTablePage } from '@/composables/useTablePage'
+import { formatTime } from '@/utils/time'
 
 const columns = [
-  {
-    title: 'id',
-    dataIndex: 'id',
-  },
-  {
-    title: '账号',
-    dataIndex: 'userAccount',
-  },
-  {
-    title: '用户名',
-    dataIndex: 'userName',
-  },
-  {
-    title: '头像',
-    dataIndex: 'userAvatar',
-  },
-  {
-    title: '简介',
-    dataIndex: 'userProfile',
-  },
-  {
-    title: '用户角色',
-    dataIndex: 'userRole',
-  },
-  {
-    title: '创建时间',
-    dataIndex: 'createTime',
-  },
-  {
-    title: '操作',
-    key: 'action',
-  },
+  { title: 'id', dataIndex: 'id' },
+  { title: '账号', dataIndex: 'userAccount' },
+  { title: '用户名', dataIndex: 'userName' },
+  { title: '头像', dataIndex: 'userAvatar' },
+  { title: '简介', dataIndex: 'userProfile' },
+  { title: '用户角色', dataIndex: 'userRole' },
+  { title: '创建时间', dataIndex: 'createTime' },
+  { title: '操作', key: 'action' },
 ]
 
-// 展示的数据
-const data = ref<API.UserVO[]>([])
-const total = ref(0)
-
-// 搜索条件
-const searchParams = reactive<API.UserQueryRequest>({
-  pageNum: 1,
-  pageSize: 10,
-})
-
-// 获取数据
-const fetchData = async () => {
-  const res = await listUserVoByPage({
-    ...searchParams,
+const { data, loading, searchParams, pagination, fetchData, doSearch, onTableChange } =
+  useTablePage<API.UserVO, API.UserQueryRequest>({
+    defaultParams: { pageNum: 1, pageSize: 10 },
+    fetchFn: async (params) => {
+      const res = await listUserVoByPage(params)
+      if (res.data.data) {
+        return res.data.data
+      }
+      message.error('获取数据失败，' + res.data.message)
+      return null
+    },
   })
-  if (res.data.data) {
-    data.value = res.data.data.records ?? []
-    total.value = res.data.data.totalRow ?? 0
-  } else {
-    message.error('获取数据失败，' + res.data.message)
-  }
-}
 
-// 分页参数
-const pagination = computed(() => {
-  return {
-    current: searchParams.pageNum ?? 1,
-    pageSize: searchParams.pageSize ?? 10,
-    total: total.value,
-    showSizeChanger: true,
-    showTotal: (total: number) => `共 ${total} 条`,
-  }
-})
-
-// 表格分页变化时的操作
-const doTableChange = (page: { current: number; pageSize: number }) => {
-  searchParams.pageNum = page.current
-  searchParams.pageSize = page.pageSize
-  fetchData()
-}
-
-// 搜索数据
-const doSearch = () => {
-  // 重置页码
-  searchParams.pageNum = 1
-  fetchData()
-}
-
-// 删除数据
-const doDelete = async (id: string) => {
-  if (!id) {
-    return
-  }
+const doDelete = async (id: number | undefined) => {
+  if (!id) return
   const res = await deleteUser({ id })
   if (res.data.code === 0) {
     message.success('删除成功')
-    // 刷新数据
     fetchData()
   } else {
     message.error('删除失败')
   }
 }
-
-// 页面加载时请求一次
-onMounted(() => {
-  fetchData()
-})
 </script>
 
 <style scoped>
