+++ yu-ai-code-mother-frontend/src/pages/user/UserRegisterPage.vue
@@ -1,3 +1,36 @@
+<script setup lang="ts">
+import { useRouter } from 'vue-router'
+import { message } from 'ant-design-vue'
+import { reactive } from 'vue'
+import { userRegister } from '@/api/userController'
+
+const router = useRouter()
+
+const formState = reactive<API.UserRegisterRequest>({
+  userAccount: '',
+  userPassword: '',
+  checkPassword: '',
+})
+
+const validateCheckPassword = (_rule: unknown, value: string, callback: (error?: Error) => void) => {
+  if (value && value !== formState.userPassword) {
+    callback(new Error('两次输入密码不一致'))
+  } else {
+    callback()
+  }
+}
+
+const handleSubmit = async (values: API.UserRegisterRequest) => {
+  const res = await userRegister(values)
+  if (res.data.code === 0) {
+    message.success('注册成功')
+    router.push({ path: '/user/login', replace: true })
+  } else {
+    message.error('注册失败，' + res.data.message)
+  }
+}
+</script>
+
 <template>
   <div id="userRegisterPage">
     <h2 class="title">鱼皮 AI 应用生成 - 用户注册</h2>
@@ -35,53 +68,6 @@
     </a-form>
   </div>
 </template>
-
-<script setup lang="ts">
-import { useRouter } from 'vue-router'
-import { userRegister } from '@/api/userController.ts'
-import { message } from 'ant-design-vue'
-import { reactive } from 'vue'
-
-const router = useRouter()
-
-const formState = reactive<API.UserRegisterRequest>({
-  userAccount: '',
-  userPassword: '',
-  checkPassword: '',
-})
-
-/**
- * 验证确认密码
- * @param rule
- * @param value
- * @param callback
- */
-const validateCheckPassword = (rule: unknown, value: string, callback: (error?: Error) => void) => {
-  if (value && value !== formState.userPassword) {
-    callback(new Error('两次输入密码不一致'))
-  } else {
-    callback()
-  }
-}
-
-/**
- * 提交表单
- * @param values
- */
-const handleSubmit = async (values: API.UserRegisterRequest) => {
-  const res = await userRegister(values)
-  // 注册成功，跳转到登录页面
-  if (res.data.code === 0) {
-    message.success('注册成功')
-    router.push({
-      path: '/user/login',
-      replace: true,
-    })
-  } else {
-    message.error('注册失败，' + res.data.message)
-  }
-}
-</script>
 
 <style scoped>
 #userRegisterPage {
