+++ yu-ai-code-mother-frontend/src/stores/loginUser.ts
@@ -1,17 +1,14 @@
 import { defineStore } from 'pinia'
 import { ref } from 'vue'
-import { getLoginUser } from '@/api/userController.ts'
+import { getLoginUser } from '@/api/userController'
 
-/**
- * 登录用户信息
- */
+const DEFAULT_LOGIN_USER: API.LoginUserVO = {
+  userName: '未登录',
+}
+
 export const useLoginUserStore = defineStore('loginUser', () => {
-  // 默认值
-  const loginUser = ref<API.LoginUserVO>({
-    userName: '未登录',
-  })
+  const loginUser = ref<API.LoginUserVO>({ ...DEFAULT_LOGIN_USER })
 
-  // 获取登录用户信息
   async function fetchLoginUser() {
     const res = await getLoginUser()
     if (res.data.code === 0 && res.data.data) {
@@ -19,9 +16,8 @@ export const useLoginUserStore = defineStore('loginUser', () => {
     }
   }
 
-  // 更新登录用户信息
-  function setLoginUser(newLoginUser: any) {
-    loginUser.value = newLoginUser
+  function setLoginUser(newLoginUser: Partial<API.LoginUserVO>) {
+    loginUser.value = { ...DEFAULT_LOGIN_USER, ...newLoginUser }
   }
 
   return { loginUser, fetchLoginUser, setLoginUser }
