+++ yu-ai-code-mother-frontend/src/types/chat.ts
@@ -1,0 +1,6 @@
+export interface ChatMessage {
+  type: 'user' | 'ai'
+  content: string
+  loading?: boolean
+  createTime?: string
+}
