+++ yu-ai-code-mother-frontend/src/types/common.ts
@@ -1,0 +1,10 @@
+export interface PageParams {
+  pageNum?: number
+  pageSize?: number
+}
+
+export interface PaginationState {
+  current: number
+  pageSize: number
+  total: number
+}
